(355) Resolve wrong paths and rename Trip component to TripItem
