+++ src/domains/trips/components/index.js
@@ -1,0 +1,1 @@
+export {default as TripItem} from './TripItem/TripItem';
